Basic mock infrastructure for Ticketing
