+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/ticketing/TicketingActivity.java
@@ -1,43 +1,33 @@
 package ch.epfl.sweng.eventmanager.ui.ticketing;
 
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
-import ch.epfl.sweng.eventmanager.repository.data.Event;
 import ch.epfl.sweng.eventmanager.repository.data.EventTicketingConfiguration;
 import ch.epfl.sweng.eventmanager.ticketing.TicketingService;
+import ch.epfl.sweng.eventmanager.ticketing.TicketingServiceManager;
 import ch.epfl.sweng.eventmanager.ui.eventSelector.EventPickingActivity;
 import ch.epfl.sweng.eventmanager.ui.eventShowcase.EventShowcaseActivity;
+import dagger.android.AndroidInjection;
+
+import javax.inject.Inject;
 
 /**
  * @author Louis Vialar
  */
 public abstract class TicketingActivity extends AppCompatActivity {
+    static final String SELECTED_EVENT_ID = "ch.epfl.sweng.SELECTED_EVENT_ID";
+    static final String TICKETING_CONFIGURATION = "ch.epfl.sweng.TICKETING_CONFIGURATION";
     private static final String TAG = "TicketingActivity";
-    public static final String SELECTED_EVENT_ID = "ch.epfl.sweng.SELECTED_EVENT_ID";
-    public static final String TICKETING_CONFIGURATION = "ch.epfl.sweng.TICKETING_CONFIGURATION";
-
+    @Inject
+    protected TicketingServiceManager manager;
+    TicketingService service;
     private int eventId;
     private EventTicketingConfiguration configuration;
-    protected TicketingService service;
 
-    public static Intent start(Event event, Context context) {
-        if (event.getTicketingConfiguration() == null) {
-            throw new UnsupportedOperationException("this event doesn't support ticketing");
-        }
-
-        TicketingService service = TicketingService.getService(event.getId(), event.getTicketingConfiguration(), context);
-
-        Intent openIntent = new Intent(context, getNextActivityForState(service));
-        openIntent.putExtra(SELECTED_EVENT_ID, event.getId());
-        openIntent.putExtra(TICKETING_CONFIGURATION, event.getTicketingConfiguration());
-        return openIntent;
-    }
-
-    protected static Class<? extends TicketingActivity> getNextActivityForState(TicketingService service) {
+    static Class<? extends TicketingActivity> getNextActivityForState(TicketingService service) {
         if (service.isLoggedIn()) {
             if (service.hasMultipleConfigurations()) {
                 return TicketingConfigurationPickerActivity.class;
@@ -49,14 +39,14 @@ public abstract class TicketingActivity extends AppCompatActivity {
         }
     }
 
-    protected Intent switchActivity(Class<? extends TicketingActivity> next) {
+    Intent switchActivity(Class<? extends TicketingActivity> next) {
         Intent openIntent = new Intent(this, next);
         openIntent.putExtra(SELECTED_EVENT_ID, eventId);
         openIntent.putExtra(TICKETING_CONFIGURATION, configuration);
         return openIntent;
     }
 
-    protected Intent backToShowcase() {
+    Intent backToShowcase() {
         Intent openIntent = new Intent(this, EventShowcaseActivity.class);
         openIntent.putExtra(EventPickingActivity.SELECTED_EVENT_ID, eventId);
         return openIntent;
@@ -64,6 +54,8 @@ public abstract class TicketingActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
+        AndroidInjection.inject(this);
+
         super.onCreate(savedInstanceState);
 
         // Get model
@@ -76,7 +68,7 @@ public abstract class TicketingActivity extends AppCompatActivity {
         } else if (configuration == null) {
             Log.e(TAG, "Got no configuration for event " + eventId);
         } else {
-            this.service = TicketingService.getService(this.eventId, this.configuration, this);
+            this.service = manager.getService(this.eventId, this.configuration, this);
         }
     }
 }
