+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/ticketing/TicketingModule.java
@@ -1,7 +1,10 @@
 package ch.epfl.sweng.eventmanager.ui.ticketing;
 
+import dagger.Binds;
 import dagger.Module;
 import dagger.android.ContributesAndroidInjector;
+
+import javax.inject.Singleton;
 
 /**
  * @author Louis Vialar
